instagram_meme_lifecycle_analysis-main/requirements.txt
python-dotenv
pandas
numpy
matplotlib
seaborn
plotly
dash
scikit-learn
snscrape
PRAW
Instaloader
spaCy
lifelines
NetworkX
